Add more experiments, auto-discovery, parallelize across runners
