--- .github/workflows/nightly_ci.yaml
@@ -1,44 +1,0 @@
-name: Nightly CI
-
-on:
-    # Allow to manually trigger the workflow
-    workflow_dispatch:
-    pull_request:
-        branches:
-            - main
-
-    # Scheduled run at 3AM PST (11AM UTC)
-    # schedule:
-    #     - cron: "0 11 * * *"
-
-jobs:
-    tests:
-        name: Nightly CI
-        runs-on:
-            - research-cluster
-        steps:
-            - name: Checkout repository
-              uses: actions/checkout@v4
-              with:
-                  submodules: true
-            - name: Install uv
-              uses: astral-sh/setup-uv@v5
-              with:
-                  enable-cache: true
-                  cache-dependency-glob: "uv.lock"
-            - name: Install dependencies
-              run: uv sync --locked
-            - name: Run tests
-              env:
-                  USERNAME_CI: CI_RUNNER
-                  WANDB_API_KEY: ${{ secrets.WANDB_API_KEY }}
-                  # Set WANDB_MODE to online only if WANDB_API_KEY is available, otherwise set to offline
-                  # This is to allow running tests on forks without WANDB_API_KEY
-                  WANDB_MODE: ${{ secrets.WANDB_API_KEY && 'online' || 'offline' }}
-                  HF_TOKEN: ${{ secrets.HF_TOKEN }}
-                  OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
-                  PRIME_API_KEY: ${{ secrets.PRIME_API_KEY }}
-                  GITHUB_REF_NAME: ${{ github.ref_name }}
-                  GITHUB_HEAD_REF: ${{ github.head_ref }}
-                  PYTEST_OUTPUT_DIR: /tmp/outputs
-              run: uv run pytest tests/nightly -m gpu
