+++ configs/ci/nightly/acereason_math.toml
@@ -1,0 +1,42 @@
+inference_gpu_ids = [0,1,2,3,4,5]
+trainer_gpu_ids = [6,7]
+
+max_steps = 500
+
+[model]
+name = "deepseek-ai/DeepSeek-R1-Distill-Qwen-7B"
+
+[ckpt] # Checkpoint at the end of training
+
+[orchestrator]
+batch_size = 1024
+seq_len = 8192
+rollouts_per_example = 8
+
+[orchestrator.sampling]
+temperature = 0.6
+max_tokens = 8192
+
+[[orchestrator.env]]
+id = "single-turn-math"
+args = { dataset_name = "nvidia/AceReason-Math", dataset_subset = "default", question_key = "problem" }
+
+[orchestrator.eval]
+interval = 50
+
+[[orchestrator.eval.env]]
+id = "math500"
+rollouts_per_example = 1
+
+[[orchestrator.eval.env]]
+id = "aime2024"
+rollouts_per_example = 32
+
+[[orchestrator.eval.env]]
+id = "aime2025"
+rollouts_per_example = 32
+
+[trainer.model.ac]
+
+[inference.parallel]
+dp = 6
